requirements.txt
dvc==3.62.0
fastapi==0.116.1
joblib==1.3.2
jupyter==1.0.0
keras==2.10.0
matplotlib==3.7.2
mlflow==3.2.0
numpy==1.26.4
optuna==4.4.0
pandas==2.3.1
pytest==8.4.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.3
tensorflow==2.10.0
uvicorn==0.35.0
wfdb==4.3.0
xgboost==3.0.4
python-dotenv==1.1.1
imblearn==0.0
requests==2.31.0
psutil==5.9.6
